Use company-box instead of company-quickhelp
While I'm not sure if this is still sufficiently maintained and is rather
bloated, company-quickhelp appears to have some bugs that sometimes display the
documentation frame in the wrong place, and that is massively irritating. Doom
still seems to use company-box, so let's hope this has enough traction in the
community.